+++ utils/logger.js
@@ -6,9 +6,11 @@
  * - 1 (Warn)
  * - 0 (Error)
  */
+var colors = require('colors/safe');
+
 var Logger = function (log_level) {
-    this._log_level = log_level ? log_level : 2
-}
+    this._log_level = log_level ? log_level : 2;
+};
 
 Logger.prototype = {
     _timestamp: function (msg) {
@@ -19,29 +21,29 @@ Logger.prototype = {
         if (this._log_level < 3) {
             return;
         }
-        console.info("[" + this._timestamp() + "] DEBG: " + msg);
+        console.info(colors.green("[" + this._timestamp() + "]"), colors.blue(" DEBG: " + msg));
     },
 
     info: function (msg) {
         if (this._log_level < 2) {
             return;
         }
-        console.info("[" + this._timestamp() + "] INFO: " + msg);
+        console.info(colors.green("[" + this._timestamp() + "]"), colors.green(" INFO: " + msg));
     },
 
     warn: function (msg) {
         if (this._log_level < 1) {
             return;
         }
-        console.warn("[" + this._timestamp() + "] WARN: " + msg);
+        console.warn(colors.green("[" + this._timestamp() + "]"), colors.yellow(" WARN: " + msg));
     },
 
     error: function (msg) {
         if (this._log_level < 0) {
             return;
         }
-        console.error("[" + this._timestamp() + "] ERRR: " + msg);
+        console.error(colors.green("[" + this._timestamp() + "]"), colors.red(" ERRR: " + msg));
     }
-}
+};
 
 exports.Logger = Logger;