+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # CHANGE LOG
 
+## v0.2.1
+
+* express 3 to 4 migration
+* user object via json file
+* added color to logger output
+* changed this.log to logger
+
 ## v0.1.6
 
 * Added this ChangeLog